pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pubmed-pipeline"
version = "0.1.0"
description = "Pipeline to query PubMed Central and check against internal datasets"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "biopython>=1.83",
  "pymed>=0.8.9",
  "pymongo>=4.6.0",
  "boto3>=1.34.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
pubmed-pipeline = "pubmed_pipeline.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
